Bugfix: NODATA=-9 is not compatible with cmask type uint8
Synop matching with npp showed this clearly

+++ read_cloudproducts_and_nwp_pps.py
@@ -345,7 +345,7 @@ def read_cma_h5(filename):
         if 'cloud_probability' in h5file.keys():
             logger.error("This CMA-file seem lika a CMAPROB file!")
     cma.cma_ext = h5file['cma_extended'].value
-    cma.cma_bin = 0*cma.cma_ext.copy()
+    cma.cma_bin = np.int64(0*cma.cma_ext.copy())
     cma.cma_bin[cma.cma_ext==1] = 1.0
     cma.cma_bin[cma.cma_ext==2] = 1.0
     cma.cma_bin[cma.cma_ext<0] =  ATRAIN_MATCH_NODATA
@@ -388,7 +388,7 @@ def read_cma_nc(filename):
         if 'cloud_probability' in pps_nc.variables.keys():
             logger.error("Probably you shold set CMAP_PROB_VALIDATION=True!")
     cma.cma_ext = pps_nc.variables['cma_extended'][0,:,:]
-    cma.cma_bin = 0*cma.cma_ext.copy()
+    cma.cma_bin = 0*np.int64(cma.cma_ext.copy())
     cma.cma_bin[cma.cma_ext==1] = 1.0
     cma.cma_bin[cma.cma_ext==2] = 1.0
     cma.cma_bin[cma.cma_ext<0] =  ATRAIN_MATCH_NODATA
